+++ LacamasFair/Views/Board/FairBoard.cshtml
@@ -22,7 +22,7 @@
 @if (isAdmin == true)
 {
     <div class="text-center">
-        <a class="btn btn-primary text-white" asp-action="AddBoardMember" asp-controller="Home">Add Board Member</a>
+        <a class="btn btn-primary text-white" asp-action="AddBoardMember" asp-controller="Board">Add Board Member</a>
     </div>
 }
